Implement version fallback and version tolerance
Version Fallback
When we as a client are supporting 1.2, but accepting a lower version, we need
to fallback to that version if the server doesn't support the best we sent.

Version Tolerance
When we as a server receive a version that is higher than what we support (like TLS 1.3 at the moment), we want to send the latest versions we *do* support.