Introduce BigIntContext to avoid allocation overhead
BigInt is now backed by BigIntStorage which can be either internally managed
or externally managed. If it is internally managed the storage uses a Swift
array. In the externally managed case the storage is taken from the current
thread's BigIntContext, which tries to minimize allocations.
It does that by allocating chunks of memory of which it hands out fragments
while the current context generation is active. When you close the current
generation, all memory is jetisoned.

This relieves the BigInt structs from the responsibility of managing their
memory, because once the generation of which they are a part is closed, all
their storage is deallocated. It also means that you can't use those BigInts
after the context has been closed.

Many of the higher-level operations that use BigInt are creating their own
context by using BigIntContext.withContext { context in ... }.
This ensures that generations aren't growing out of bounds.

Currently memory is allocated in chunks of 16k under the assumption that
each BigInt needs much less memory than this. If you have BigInts that
exceed 16k this approach fails and might need to adjust dynamically.